requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
pymupdf==1.24.9
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4
rank-bm25==0.2.2
tavily-python